First pass at cleanup for bdeverify
Currently there are still around 6400 warnings, although a large number
of these should resolve as false positives in a new iteration of
bdeverify in a few days.  Another large set are repeated warnings
about a lack of documentation each time a template is instantiated
and while these must be addressed, it will be easier to see and
address such issues once the false-positives are eliminated.