Further bde_verify updates for hash tables
This takes a proper attempt at cleaning up the bde_verify issues associated
with trying to flag allocator usage for a nested class within a class
template.

This fix moves the nested class outside the primary class template, which
solves a number of isses.  Ths documentation is now more conventional,
the nested trait problems go away, and we no longer need to work around
the MSVC bug that did not recognize the out-of-line definitions of the
methods of a nested class (in a class template).